fix mini bug on dynamics
Former-commit-id: 6b7f676936a872f5eeb28efddeeeaf7b0a5cd8c6

+++ musif/extract/features/dynamics/handler.py
@@ -136,7 +136,7 @@ def update_part_objects(
     part_features.update(
         {
             DYNMEAN: mean(dynamics) if len(dynamics) != 0 else 0,
-            DYNMEAN_WEIGHTED: dyn_mean_weighted / total_sounding_beats
+            DYNMEAN_WEIGHTED: float(dyn_mean_weighted / total_sounding_beats)
             if total_sounding_beats != 0
             else 0,
             DYNGRAD: float(dyn_grad / (len(dynamics) - 1))
